+++ src/components/common/VideosList.js
@@ -9,12 +9,12 @@ const VideosList = ({ splitScreen, videos }) => {
     return (
         <div
             className={clsx(
-                'w-full h-auto overflow-x-scroll flex md:grid gap-8 mt-8',
+                'w-full h-auto overflow-x-scroll flex md:grid gap-y-2 gap-x-8 mt-8',
                 splitScreen ? 'md:grid-cols-1 lg:grid-cols-2' : 'md:grid-cols-2 md:h-auto md:overflow-hidden'
             )}
         >
             {videos.map((props) => (
-                <VideoCard key={props?.title} className={splitScreen ? 'md:w-60' : ''} {...props} />
+                <VideoCard key={props?.title} {...props} />
             ))}
         </div>
     );